+++ src/shared/components/P/types.ts
@@ -1,0 +1,14 @@
+import { ComponentProps } from 'react';
+
+import { StyledP } from '@/shared/components/P/P.style';
+
+export interface PProps extends ComponentProps<typeof StyledP> {
+    /**
+     * If the text should appear bold
+     */
+    bold?: boolean;
+    /**
+     * If the text should be a secondary color
+     */
+    secondary?: boolean;
+}
