+++ src/shared/components/Text/Text.tsx
@@ -1,27 +1,11 @@
-import React, { useId } from 'react';
+import { useId } from 'react';
 
 import {
     StyledDescription,
     StyledInput,
     StyledLabel,
 } from '@/shared/components/Text/Text.style';
-
-export type TextTypes = 'text' | 'password' | 'email';
-
-export interface TextProps extends React.ComponentProps<typeof StyledInput> {
-    /**
-     * What type of value the input expects
-     */
-    type?: TextTypes;
-    /**
-     * Optional label placed above the input
-     */
-    label?: string;
-    /**
-     * Optional error message underneath the field, also highlights the input
-     */
-    error?: string;
-}
+import { TextProps } from '@/shared/components/Text/types';
 
 /**
  * A UI component which accepts user text input
